+++ packages/amplify-e2e-core/src/categories/custom.ts
@@ -14,29 +14,26 @@ export const addCDKCustomResource = async (cwd: string, settings: any): Promise<
     .runAsync();
 };
 
-export const addCFNCustomResource = async (cwd: string, settings: any): Promise<void> => {
-  await spawn(getCLIPath(), ['add', 'custom'], { cwd, stripColors: true })
+export const addCFNCustomResource = async (cwd: string, settings: any, testingWithLatestCodebase = false): Promise<void> => {
+  const chain = spawn(getCLIPath(testingWithLatestCodebase), ['add', 'custom'], { cwd, stripColors: true })
     .wait('How do you want to define this custom resource?')
     .send(KEY_DOWN_ARROW)
     .sendCarriageReturn()
     .wait('Provide a name for your custom resource')
     .sendLine(settings.name || '\r')
     .wait('Do you want to access Amplify generated resources in your custom CloudFormation file?')
-    .sendYes()
-    .wait('Select the categories you want this custom resource to have access to')
-    .sendCtrlA()
-    .sendCarriageReturn()
-    .wait('Do you want to edit the CloudFormation stack now?')
-    .sendNo()
-    .sendEof()
-    .runAsync();
+    .sendYes();
+  if (settings.promptForCategorySelection) {
+    chain.wait('Select the categories you want this custom resource to have access to').sendCtrlA().sendCarriageReturn();
+  }
+  await chain.wait('Do you want to edit the CloudFormation stack now?').sendNo().sendEof().runAsync();
 };
 
-export function buildCustomResources(cwd: string) {
+export function buildCustomResources(cwd: string, usingLatestCodebase = false) {
   return new Promise((resolve, reject) => {
     const args = ['custom', 'build'];
 
-    spawn(getCLIPath(), args, { cwd, stripColors: true })
+    spawn(getCLIPath(usingLatestCodebase), args, { cwd, stripColors: true })
       .sendEof()
       .run((err: Error) => {
         if (!err) {