+++ packages/amplify-e2e-tests/src/__tests__/migration/node.function.test.ts
@@ -57,7 +57,7 @@ describe('nodejs version migration tests', () => {
     );
     let authStackContent = fs.readFileSync(authStackFileName).toString();
 
-    authStackContent = authStackContent.replace('nodejs14.x', 'nodejs10.x');
+    authStackContent = authStackContent.replace('nodejs16.x', 'nodejs10.x');
 
     fs.writeFileSync(authStackFileName, authStackContent, 'utf-8');
 
@@ -72,7 +72,7 @@ describe('nodejs version migration tests', () => {
     );
     let functionStackContent = fs.readFileSync(functionStackFileName).toString();
 
-    functionStackContent = functionStackContent.replace('nodejs14.x', 'nodejs10.x');
+    functionStackContent = functionStackContent.replace('nodejs16.x', 'nodejs10.x');
 
     fs.writeFileSync(functionStackFileName, functionStackContent, 'utf-8');
 
@@ -84,8 +84,8 @@ describe('nodejs version migration tests', () => {
     functionStackContent = fs.readFileSync(functionStackFileName).toString();
 
     expect(projectConfigContent.indexOf('3.1')).toBeGreaterThan(0);
-    expect(authStackContent.indexOf('nodejs14.x')).toBeGreaterThan(0);
-    expect(functionStackContent.indexOf('nodejs14.x')).toBeGreaterThan(0);
+    expect(authStackContent.indexOf('nodejs16.x')).toBeGreaterThan(0);
+    expect(functionStackContent.indexOf('nodejs16.x')).toBeGreaterThan(0);
   });
 
   const amplifyNodeMigrationAndPush = async (cwd: string): Promise<void> => {