+++ packages/amplify-provider-awscloudformation/src/root-stack-builder/root-stack-transform.ts
@@ -1,5 +1,5 @@
-import { AmplifyRootStackTemplate } from '@aws-amplify/cli-extensibility-helper';
-import * as cdk from '@aws-cdk/core';
+import { AmplifyRootStackTemplate, getProjectInfo } from '@aws-amplify/cli-extensibility-helper';
+import * as cdk from 'aws-cdk-lib';
 import {
   $TSContext,
   AmplifyError,
@@ -78,8 +78,9 @@ export class AmplifyRootStackTransform {
           external: true,
         },
       });
+      const projectInfo = getProjectInfo();
       try {
-        await sandboxNode.run(overrideCode).override(this._rootTemplateObj as AmplifyRootStackTemplate);
+        await sandboxNode.run(overrideCode).override(this._rootTemplateObj as AmplifyRootStackTemplate, projectInfo);
       } catch (err) {
         throw new AmplifyError(
           'InvalidOverrideError',